test: 유저 검색 기능 test 코드 수정
