+++ src/test/java/com/ssafy/questory/service/FriendServiceTest.java
@@ -5,7 +5,9 @@ import com.ssafy.questory.domain.Friend;
 import com.ssafy.questory.domain.Member;
 import com.ssafy.questory.dto.request.friend.FriendStatusRequestDto;
 import com.ssafy.questory.dto.request.member.MemberEmailRequestDto;
+import com.ssafy.questory.dto.request.member.MemberSearchRequestDto;
 import com.ssafy.questory.dto.response.friend.FollowResponseDto;
+import com.ssafy.questory.dto.response.member.MemberSearchResponseDto;
 import com.ssafy.questory.dto.response.member.auth.MemberInfoResponseDto;
 import com.ssafy.questory.repository.FriendRepository;
 import com.ssafy.questory.repository.MemberRepository;
@@ -20,6 +22,7 @@ import java.util.List;
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.*;
+import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.*;
 
 class FriendServiceTest {
@@ -144,4 +147,57 @@ class FriendServiceTest {
         verify(friendRepository).findFollowRequestsByRequesterEmailWithPaging("user1@example.com", 0, 10);
         verify(friendRepository).countFollowRequestsByRequesterEmail("user1@example.com");
     }
+
+    @Test
+    @DisplayName("이메일 키워드로 유저 검색 - 페이징 포함")
+    void searchMembersByEmail_withPaging() {
+        // given
+        String keyword = "kim";
+        int page = 0;
+        int size = 2;
+        int offset = page * size;
+
+        Member member1 = Member.builder()
+                .email("kim1@example.com")
+                .nickname("김하나")
+                .profileUrl("https://cdn.questory.com/kim1.jpg")
+                .build();
+
+        Member member2 = Member.builder()
+                .email("kim2@example.com")
+                .nickname("김둘")
+                .profileUrl("https://cdn.questory.com/kim2.jpg")
+                .build();
+
+        List<Member> memberList = List.of(member1, member2);
+
+        Member requester = Member.builder()
+                .email("test@example.com")
+                .nickname("nick")
+                .password("pw")
+                .build();
+
+        // ✅ requester.getEmail() 인자 추가
+        given(memberRepository.searchByEmailWithPaging(keyword, requester.getEmail(), offset, size))
+                .willReturn(memberList);
+
+        given(memberRepository.countByEmail(keyword))
+                .willReturn(10);
+
+        MemberSearchRequestDto requestDto = MemberSearchRequestDto.builder()
+                .email(keyword)
+                .page(page)
+                .size(size)
+                .build();
+
+        // when
+        Page<MemberSearchResponseDto> result = friendService.search(requester, requestDto);
+
+        // then
+        assertThat(result.getContent()).hasSize(2);
+        assertThat(result.getTotalElements()).isEqualTo(10);
+        assertThat(result.getContent().get(0).getEmail()).isEqualTo("kim1@example.com");
+        assertThat(result.getContent().get(1).getNickname()).isEqualTo("김둘");
+    }
+
 }
